Enable CONFIG_IPV6_SUBTREES and CONFIG_NF_LOG_IPV6. JB#62994
From kernelconfig.io:

"The destination address is still the primary routing key, so mixing
 normal and source prefix specific routes in the same routing table
 may sometimes lead to unintended routing behavior. This can be
 avoided by defining different routing tables for the normal and
 source prefix specific routes."

Thus, it is good to have this enabled as well for IPv6 functioning on
the same level as on other devices. Enabled on every other device.

Also, does not hurt to have the logging option enabled on IPv6 as well.
Enabled on every other device.